requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.13.1
dash==2.9.3
flask==2.2.3
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
scipy==1.10.1
statsmodels==0.13.5
nltk==3.8.1
spacy==3.5.2
transformers==4.28.1
yfinance==0.2.18
python-dotenv==1.0.0
pytest==7.3.1
jinja2==3.1.2
weasyprint==59.0
pdfrw==0.4
ratelimit==2.2.1
requests==2.28.2
beautifulsoup4==4.11.2
apscheduler==3.10.1
pyjwt==2.6.0
bcrypt==4.0.1
passlib==1.7.4
news-api-client==0.2.7
tqdm==4.65.0
